+++ src/__tests__/reducers/app/layout.test.js
@@ -6,6 +6,7 @@ describe ('Layout reducer', () => {
     let initialState = {
         isLoading: false,
         isNavbarClosed: false,
+        items: [],
     };
 
     it ('Test default state returns default values', () => {