+++ src/__tests__/views/DashboardPage/DashboardPage.test.js
@@ -1,13 +1,33 @@
 import { render, screen} from '@testing-library/react';
+import configureStore from "redux-mock-store";
+import thunk from "redux-thunk";
 import DashboardPage from "../../../views/DashboardPage/DashboardPage";
+
+const middlewares = [thunk];
+const mockStore = configureStore(middlewares);
+
+const initialState = {
+    balance: {
+        items: {
+            balance: "1234.56"
+        }
+    }
+};
+let store = mockStore(initialState);
 
 describe('Test DashboardPage', () => {
     it('Can render DashboardPage', () => {
-
-        render(<DashboardPage />);
-
+        render(<DashboardPage store={store} />);
         const title = screen.getByText(/dashboard/i);
         expect(title).toBeInTheDocument();
+    });
+
+    it('Can render Kpi balance', () => {
+        render(<DashboardPage store={store} />);
+        const title = screen.getByText(/balance/i);
+        expect(title).toBeInTheDocument();
+        const balance = screen.getByText(/1234.56/i);
+        expect(balance).toBeInTheDocument();
 
     });
-});
+})