+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Add CircleCI
 + Add balance api and reducer
++ Add KPI
 
 ## 0.1.0
 + First commit