+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Add CircleCI
 + Add balance api and reducer
 + Add KPI
++ Add KPI balance (fake)
 
 ## 0.1.0
 + First commit