+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Add balance api and reducer
 + Add KPI
 + Add KPI balance (fake)
++ Add balance (real)
 
 ## 0.1.0
 + First commit