+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-# 0.1.0
+# Next version
++ Add CircleCI
++ Add balance api and reducer
++ Add KPI
++ Add KPI balance (fake)
++ Add balance (real)
+
+## 0.1.0
 + First commit